requirements.txt
fastapi
python-multipart
mysql-connector-python
uvicorn